restartable. Change doLaunchP default from restartP to startP .
Because it's the only safe action: if program already running (like browser),
restartP will terminate and restart it at xmonad restart, which is probably
not what user wants. On the other hand, startP just starts it, if it's not
already running and do nothing otherwise.

Thus, for all docks, which i want to restart with xmonad, i need to explicitly
set doLaunchP to restartP.