restartable. `serialize` returns arguments in MonadIO .
This allows more powerful argument's processing.

Xmobar now is able to normalize config path: if only filename is specified,
it's assumed to be relative to home directory and is converted into absolute
path. Otherwise, path preserved as is.

This replaces config existence check in Xmobar's `runP`, which leads to
incorrect results, when xmonad is launched not from home directory and xmobar
config exists: xmonad thinks, that config exists (and it really is, in home
directory), but xmobar can't find it, because it looks relative to current
directory.

And because xmobar config is used in Eq instance and determining home
directory requires IO, i can't normalise it in other places (like Lens or Eq
instance), except `serialise`.